requirements.txt
streamlit>=1.28.0
marker-pdf[full]>=1.8.0
torch>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
psutil>=5.9.0